Add a SuccessfulSeparation interface
If an instance of Separation implements this interface, then this means
that all separation problems are guaranteed to be solvable. This can
then be used to speed up Petri net synthesis, because the individual
separation problems do not have to be checked one by one.

MarkedGraphSeparation implements this new interface.

This commit speeds up synthesis of the reachability graph of a cycle of
size 100 with two tokens. Previously, this took about ten seconds. With
this change it only takes about one seconds. Both times are with
assertions disabled.