Synthesis: Introduce a Synthesizer interface
SynthesizePN no longer does the "real job". It's just a class that is
supposed to make other things easier to use. Instead, a Synthesizer is
added that does all of the work. Right now, this is only meant to move
around some code, but this new interface will be used in later commits.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>