src/ant: Remove UnparsableException
Once upon a time, parsers threw all sorts of weird exceptions and thus
we needed special handling for "this parser is expected to throw this
kind of exception". For a while now we have a ParseException and parsers
are expected to only throw this if a file cannot be parsed. Thus,
UnparsableException is no longer needed.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>