PetriNetOrTransitionSystem: Add coverability graph accessors
Most uses of PetriNetOrTransitionSystem really do want an LTS. If a
Petri net is given, its reachability graph is computed and that one is
used.

Simplify this by adding member functions to PetriNetOrTransitionSystem
that to exactly this: Either return the "stored" transition system or
compute the reachability/coverability graph of the stored Petri net.

Since this adds a dependency onto the CoverabilityGraph class,
PetriNetOrTransitionSystem has to be moved from src/lib to src/module.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>