Bisimulation: Only accept transition systems
Bisimulation does not have anything to do with Petri nets. If one was
given, it would compute the Petri net's reachability graph and use that
instead. This should be done by the caller instead, I think.

Thus, this commit changes the parameters from PetriNetOrTransitionSystem
to TransitionSystem and changes the calling code suitably.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>