Marking: Inline some easy cases of setMarking(Marking)
In the cases that this commit touches, it is guaranteed that this.net ==
m.net and thus the code in setMarking() would simply copy over some
data. This can be done directly.

In the case of Marking#fire(), a copy can even be avoided: Since the
temporary Marking "result" becomes unreachable, we can share its token
list and do not have to copy it.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>