tests: parsers: fix/add testMissingNewlineAfterComment
Commit 82ca5a74be96e56e876b45c71a7ed4408faf5fc1 took tests from
APTParserTest (which tested the old antlr3 based parser) and reused them
into the new AptPNParserTest. This conversion erroneously changed the
expected result of testMissingNewlineAfterComment. Thus, we were testing
for the existence of a bug. This commit fixes the expectation.

The AptLTSParserTest had no corresponding test, this commits adds one.

This leads to two test failures. The next commit will fix them.

Signed-off-by: vsp <vsp@chuck.Informatik.Uni-Oldenburg.DE>