Improvements to Algorithm for Invariants "PIPE"
Refactor code for phase 2 of PIPE algorithm to work with a *set* of
normalized columns. This avoids heavy construction/destruction costs
from working with a matrix stored in row format to add/remove columns
repeatedly. 
Also remove columns which have become useless from further consideration
during computation.

Improvements are integrated in our forked version of this at :
https://github.com/lip6/ITSTools/tree/master/smt/fr.lip6.move.gal.gal2smt/src/uniol/apt/analysis/invariants
This commit is to feedback the nice improvements back to apt group.
thanks for making that code public in the first place.