OverapproximatePN: Turn a check into an assertion
For k-bounded regions, this was checking if the region of the lts before
expansion is no longer a k-bounded region of the lts after expansion.
However, this necessarily holds always: When a new arc+state is added,
then there was an unsolvable ESSP instance. If we have a region which is
no longer valid afterwards, its complementary region would solve the
ESSP instance.

Thus, turn this check into an assertion (because assertions are exactly
meant for things that must always be true).

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>