Optimise AllSmallCyclesHavePVOne
Harro found out that phase 2 of the existing algorithm is almost enough
to check everything. This commit changes phase 2 so that instead of just
verifying that all chords have Parikh vectors where all entries are the
same, phase 2 now checks that all chords have Parikh vector of all-ones.

With this change, phase 1 becomes obsolete. Proof sketch is in inline
comments in code.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>