Add a faster implementation for finding smallest cycles
With this new implementation, computing the smallest cycle of the
reachability graph of a 10 bit net takes about a tenth of a second. The
old algorithm needs 26 seconds for the same task.

Similar, for the reachability graph of 8 connected bit nets, the new
algorithm takes less than 2/100th of a second while the old algorithm
needs about 9 seconds.

This new algorithm is not always applicable. For example, the
reachability graph of 7 tristate philosophers is not persistent and thus
the new algorithm cannot be used. Here, the new algorithm just needs
1/1000th of a second to fail. Compared with the 30 seconds that the
existing algorithm needs, this has barely any negative effect on the
total running time.

Thus, this commit also changes ComputeSmallestCycles to always try to
use this new implementation and only fall back to the existing one if
the new implementation throws a PreconditionFailedException.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>