ComputeSmallestCycles: Speed up special case for disjoint cycles
The algorithm behind computePVsOfSmallestCycles() can fail in different
ways. Either one of its general preconditions is not satisfied, or its
special precondition of "disjoint small cycles" is not satisfied. The
later is just what checkSameOrMutallyDisjointPVs() wants to check for.

Thus, use the exception that is thrown in that case to figure out the
result of the check.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>