Speed up getSource()/getTarget() on edges
Up to now, an Edge remembers its source and target as IDs. This commit
changes the code so that instead the object is used directly. This means
that getSource() and getTarget() now work in constant time instead of
logarithmic time when a Map lookup was needed.

I am not totally sure on the effects of this change for the code in APT,
but something that I am currently working on only needs half as much
time after this change.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>