TransitionSystemTest: Add a (failing) test
A TransitionSystem saves the post- and preset nodes of a state as a Set.
When an Arc is removed, it (wrongly) assumes that these set needs to be
updated, because there is no more connection between these nodes.
However, this is wrong, because there could be multiple Arcs connecting
two given nodes.

Thanks to Judith Segerath for finding this problem.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>