TransitionSystem: Use Bags for pre- and postset nodes
Instead of saving pre- and postset nodes as Sets, this now uses Bags
from Apache Commons Collections. A Bag can contain a given item multiple
times and thus this implicitly counts how often a given state is in the
pre/postset of another state.

This fixes the test that was added in the previous commit.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>