Add new SynthesisFactorisation class
The existing Factorization class is nice, but it is too slow. This new
class does not do proper factorization in general, but only if its input
is Petri net solvable. If the input is not Petri net solvable, the
guarantee is that one of the computed factors also is not Petri net
solvable.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>