CycleSearchViaChords: Compute shorter paths
For each chord going from state s via label a to state r, this code
computes PV(s)+a-PV(r), i.e. the difference of the Parikh vectors of the
two states (plus one a). When the paths from the initial state to s/r
get longer, it becomes more expensive to compute PV(s) and PV(r).

This commit speeds things up by computing a common ancestor of s and r
and computing the Parikh vectors from this ancestor to s and r, instead
of from the initial state to s and r. This is valid since the code
computes the difference of the two Parikh vectors and we are removing
the same amount from both vectors.

The following table shows the effect of this change when computing the
cycles of connected bit nets, as generated by APT's net generator.

size   time before   time after
 14         2s           2s
 15         5s           4s
 16        10s           8s
 17        24s          19s
 18        60s          45s

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>