PetriNet: Test copy constuctor after a place was removed
This adds a new test to PetriNetTest which creates a place, gives that
place an initial token count, removes the place and finally tries to
copy the Petri net.

Currently, this will throw an extension since the constructor
Marking(PetriNet, Marking) will try to find the copy of the already
removed place and end up with an exception.

This is broken since commit 2b05f6f9574f2 which removed a call to
ensureConsistency() at the beginning of setMarking() (which later in
commit ccc71fd08560ea9 was inlined into this constructor).

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>