ElementarySeparation: Use Lists<A> instead of Map<String, A> for labels
Labels are already assigned an index by RegionUtility. Make use in this
in ElementarySeparation so that this class causes less memory
allocations (always good; fewer HashMap entries are allocated now) and
is hopefully a bit faster (the copy constructor of RoughRegion is one of
the hot spots of the code).

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>