ElementarySeparation: Use an array-based class instead of HashMap
A HashMap consists of awfully many objects that are slow to copy. This
new class instead uses two arrays: One array containing states (thus
assigning indicies to states) and another array containing Boolean
(indicating if the state is in/out/unknown the region).

This gets rid of the copy constructor of RoughRegion as a hot spot.
Instead, the binary search that is used to figure out the index of a
state is now a hot spot...

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>