ElementarySeparation: Add an event order heuristics
For this algorithm, the best case is that the already assigned states
are all close to each other in the lts. However, when we do not know how
to continue, we randomly pick a label and assign it some operations.
Thus, nothing tries to assign operations to nearby states. (The random
order is actually alphabetic, which is still bad.)

This commit improves that by adding an event order. Events are assigned
operations based on this order. The order here is not alphabetic, but
based on a depth-first search through the lts. Thus, events which are
nearby to each other are more likely to get assigned an operation right
after each other.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>