Add streaming variant of ParameterTransformation
When reading in a big LTS file, the file is first read into a string, so
decoding the utf8 in the file. This string is then re-encoded into a
byte string by the existing parser mechanism
(AbstractParser#parseString(String)). This is both unnecessary and
wasteful, plus it can cause issues:

When encoding a string, the encoder makes a worst case assumption that
every character will need the maximum number of bytes. This means that
for ASCII files with more than 1 GiB size, the resulting string
overflows and encoding fails.

Fix this by just not decoding & encoding the data: A special
StreamParameterTransformation class is introduced that allows to handle
an InputStream instead of a String. The mechanism in
ParametersTransformerImpl recognises this class and passes in the stream
directly, if one is available.

Note that this does not yet mean that APT can actually parse huge LTS
files, because ANTLR tries to allocate a huge array internally, too,
which makes parsing fail again due to an overflow.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>