SMTInterpolHelper: Simplify requireConflictFree()
Manuel noticed that, thanks to plainness, expressing "at most a single
transition consumes tokens" is easy: The sum of the backwards weights
may be at most one.

This is a lot simpler than the previous approach that created a big
disjunction saying "either the sum of all but the first event is zero,
or the sum of all but the second event is zero, or...".

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>