Handle absolute window paths specially in ParametersTransformerImpl
Thanks to the option support that recently handled, we can now e.g.
directly read lola files via lola:name_of_the_file.lola. However, this
interferes with absolute files on Windows: The "C" in "C:\Something" is
interpreted as an option instead of part of the file path.

Handle this by detecting this case and not handling things as an option
in this case.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>